Fix MappingIsoparametric 1D scaling, add another test
