+++ tests/test_assembly.py
@@ -466,105 +466,5 @@ class TestThreadedAssembly(TestCase):
         )
 
 
-@pytest.mark.parametrize(
-    "m,mdgtype,etype,check1,check2",
-    [
-        (
-            MeshTri.init_tensor(np.linspace(0, 1, 7),
-                                np.linspace(0, 1, 7)),
-            MeshTri1DG,
-            ElementTriP1,
-            lambda x: x[0] == 1,
-            lambda x: x[0] == 0,
-        ),
-        (
-            MeshTri.init_tensor(np.linspace(0, 1, 5),
-                                np.linspace(0, 1, 5)),
-            MeshTri1DG,
-            ElementTriP1,
-            lambda x: x[0] == 1,
-            lambda x: x[0] == 0,
-        ),
-        (
-            MeshTri().refined(2),
-            MeshTri1DG,
-            ElementTriP1,
-            lambda x: x[0] == 1,
-            lambda x: x[0] == 0,
-        ),
-        (
-            MeshTri().refined(3),
-            MeshTri1DG,
-            ElementTriP1,
-            lambda x: x[0] == 1,
-            lambda x: x[0] == 0,
-        ),
-        (
-            MeshQuad().refined(2),
-            MeshQuad1DG,
-            ElementQuad1,
-            lambda x: x[0] == 1,
-            lambda x: x[0] == 0,
-        ),
-        (
-            MeshQuad().refined(2),
-            MeshQuad1DG,
-            ElementQuad2,
-            lambda x: x[0] == 1,
-            lambda x: x[0] == 0,
-        ),
-        (
-            MeshTri().refined(2),
-            MeshTri1DG,
-            ElementTriP2,
-            lambda x: x[0] == 0,
-            lambda x: x[0] == 1,
-        ),
-        (
-            MeshTri().refined(2),
-            MeshTri1DG,
-            ElementTriP2,
-            lambda x: x[1] == 0,
-            lambda x: x[1] == 1,
-        ),
-        (
-            MeshHex().refined(4),
-            MeshHex1DG,
-            ElementHex1,
-            lambda x: x[1] == 0,
-            lambda x: x[1] == 1,
-        ),
-        (
-            MeshHex().refined(3),
-            MeshHex1DG,
-            ElementHex2,
-            lambda x: x[2] == 0,
-            lambda x: x[2] == 1,
-        ),
-    ]
-)
-def test_periodic_mesh_assembly(m, mdgtype, etype, check1, check2):
-
-    def _sort(ix):
-        # sort index arrays so that ix[0] matches
-        return ix[np.argsort(np.sum(m.p, axis=0)[ix])]
-
-    mp = mdgtype.periodic(m,
-                          _sort(m.nodes_satisfying(check1)),
-                          _sort(m.nodes_satisfying(check2)))
-    basis = Basis(mp, etype())
-    A = laplace.assemble(basis)
-
-    f = unit_load.assemble(basis)
-    D = basis.get_dofs()
-    x = solve(*condense(A, f, D=D))
-
-    if m.dim() == 2:
-        assert_almost_equal(x.max(), 0.125)
-    else:
-        assert_almost_equal(x.max(), 0.07389930610869411, decimal=3)
-    assert not np.isnan(x).any()
-
-
 if __name__ == '__main__':
     main()
